vistas entre empleados y clientes

+++ src/main/java/cl/m5d12/dao/Checklist.java
@@ -15,7 +15,7 @@ public class Checklist implements Serializable {
 
 	@Id
 	@Column(name="ID_CHECKLIST")
-	private long idChecklist;
+	private int idChecklist;
 
 	@Column(name="DETALLE_CHEQUEO")
 	private String detalleChequeo;
@@ -31,14 +31,14 @@ public class Checklist implements Serializable {
 		
 	}
 	
-	public Checklist(long idChecklist, String detalleChequeo, int idVisitas) {
+	public Checklist(int idChecklist, String detalleChequeo, int idVisitas) {
 		super();
 		this.idChecklist = idChecklist;
 		this.detalleChequeo = detalleChequeo;
 		this.idVisitas = idVisitas;
 	}
 
-	public Checklist(long idChecklist, String detalleChequeo, String estado, int idVisitas) {
+	public Checklist(int idChecklist, String detalleChequeo, String estado, int idVisitas) {
 		super();
 		this.idChecklist = idChecklist;
 		this.detalleChequeo = detalleChequeo;
@@ -46,11 +46,11 @@ public class Checklist implements Serializable {
 		this.idVisitas = idVisitas;
 	}
 
-	public long getIdChecklist() {
+	public int getIdChecklist() {
 		return idChecklist;
 	}
 
-	public void setIdChecklist(long idChecklist) {
+	public void setIdChecklist(int idChecklist) {
 		this.idChecklist = idChecklist;
 	}
